feat: Refactor model loading hooks and update import paths for better organization
